pyproject.toml
[project]
name = "openvoice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    'librosa==0.9.1',
    'faster-whisper==0.9.0',
    'pydub==0.25.1',
    'wavmark==0.0.3',
    'numpy==1.22.0',
    'eng_to_ipa==0.0.2',
    'inflect==7.0.0',
    'unidecode==1.3.7',
    'whisper-timestamped==1.14.2',
    'pypinyin==0.50.0',
    'cn2an==0.5.22',
    'jieba==0.42.1',
    'gradio==3.48.0',
    'langid==1.1.6',
    'onnxruntime==1.19.2',
    "melotts",
    "botocore==1.34.88",
    "cached_path==1.6.2",
    "fastapi==0.115.5",
    "pydantic==2.10.2",
    "termcolor==2.5.0",
    "path-py==12.5.0",
]
package = true
[tool.uv]
dev-dependencies = [
    "notebook==7.2.2",
]

[tool.uv.sources]
melotts = { git = "https://github.com/myshell-ai/MeloTTS.git", rev = "5b53848" }
